Update STP_plots.py
Note: A change was made to the KEY lat/lon so that it would be within the meso domain. Old coords were 24.553, -81.788. New coords are 24.722, -81.390.

Also updated some plots and added a filter to make sure anoumalously low mesoanalysis values are set to zero (until I figure out whether or not a mask is in place).