Expanded Alpaca functionality.
Expanded the Alpaca functionality, including the ability to buy/sell stocks. Created different versions of the SVR forecast routine for deployment and training. Included list of 100 stocks to start with. Need to incorporate SQL database to keep track of when a stock was last bought/sold. Also need to add in functionality for sending a report of transactions and perhaps a way to plot data online?